% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_acct.R
\name{format_acct}
\alias{format_acct}
\title{Function to simply format currency values
Formatting will follow Excel Accounting style}
\usage{
format_acct(x, digits = 2, shown.in = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{numeric to denote number of significant digits}

\item{shown.in}{character to express value in Thousands, Millions, or Billions}
}
\value{
character vector of length(x) with desired formatting
}
\description{
Function to simply format currency values
Formatting will follow Excel Accounting style
}
